Group alternative benchmarks
We want to see how everything compares together.
